backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
setuptools<81
python-multipart==0.0.6
pymupdf==1.23.8
python-pptx==0.6.23
google-generativeai==0.3.2
openai-whisper==20231117
imageio-ffmpeg==0.5.1
httpx==0.26.0
python-docx==1.1.0
pydantic-settings==2.1.0
aiosqlite==0.19.0
